+++ CHANGES.rst
@@ -16,6 +16,9 @@ Unreleased
   :class:`.LiveThreadContribution`.
 * :meth:`.LiveThreadContribution.add` to add an update to the live thread.
 * :meth:`.LiveThreadContribution.close` to close the live thread permanently.
+* :attr:`.LiveUpdate.contrib` to obtain an instance of
+  :class:`.LiveUpdateContribution`.
+* :meth:`.LiveUpdateContribution.remove` to remove a live update.
 
 **Fixed**